--- tests/_support/ProjectTestCase.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Tests\Support;
-
-use CodeIgniter\Test\CIUnitTestCase;
-
-/**
- * @internal
- */
-abstract class ProjectTestCase extends CIUnitTestCase
-{
-    use TestCaseTrait;
-}
